Command_Control,_Exfiltration_&_Persistance/Empire/Empire/pyproject.toml
[tool.poetry]
name = "empire-bc-security-fork"
version = "4.2.0"
description = ""
authors = ["BC Security <info@bc-security.org>"]
readme = "README.md"
homepage = "https://github.com/BC-SECURITY/Empire"
repository = "https://github.com/BC-SECURITY/Empire.git"
packages = [
    { include = "empire.py" },
    { include = "empire/server/**/*.py" },
    { include = "empire/client/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.8"
urllib3 = "*"
requests = "^2.24.0"
setuptools = "*"
iptools = "*"
pydispatcher = "*"
flask = "^1.1.2"
macholib = "*"
dropbox = "*"
pyOpenSSL = "*"
pyinstaller = "*"
zlib_wrapper = "*"
netifaces = "*"
jinja2 = "*"
xlutils = "*"
pefile = "*"
bcrypt = "*"
pyparsing = "*"
PyMySQL = "^0.10.1"
SQLAlchemy = "^1.3.20"
PyYAML = "^5.3.1"
SQLAlchemy-Utc = "^0.11.0"
prompt-toolkit = "^3.0.9"
terminaltables = "^3.1.0"
docopt = "^0.6.2"
humanize = "^3.2.0"
pydantic = "^1.8.1"
pycryptodome = "^3.10.1"
cryptography = "^3.4.7"
Flask-SocketIO = "^5.0.1"
python-socketio = "^5.2.1"
jq = "^1.1.3"
donut-shellcode = "^0.9.2"
websockify = "^0.10.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
